requirements.txt
langchain
openai
snscrape
python-dotenv
notion-client
